requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.1.1
numpy==2.3.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.3.0
protobuf==6.31.1
pyarrow==21.0.0
pybind11==3.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tflite-support==0.1.0a1
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xgboost==3.0.4
